chore: rearrange tests
